Refactor player input handling
