+++ Assets/Scripts/PlayerController.cs
@@ -1,7 +1,7 @@
 using UnityEngine;
-using UnityEngine.InputSystem;
 
 [RequireComponent(typeof(Rigidbody2D))]
+[RequireComponent(typeof(PlayerInputController))]
 public class PlayerController : MonoBehaviour
 {
     [SerializeField] private float moveSpeed = 5f;
@@ -9,35 +9,25 @@ public class PlayerController : MonoBehaviour
     [SerializeField] private Transform groundCheck;
     [SerializeField] private float groundCheckRadius = 0.2f;
     [SerializeField] private LayerMask groundLayer;
+
     private Rigidbody2D _rigidbody;
-    private float _moveInput;
-    private bool _jumpRequested;
+    private PlayerInputController _inputController;
 
     private void Awake()
     {
         _rigidbody = GetComponent<Rigidbody2D>();
-    }
-
-    private void Update()
-    {
-        _moveInput = ReadHorizontalInput();
-
-        if (IsJumpTriggered() && IsGrounded())
-        {
-            _jumpRequested = true;
-        }
-
+        _inputController = GetComponent<PlayerInputController>();
     }
 
     private void FixedUpdate()
     {
         Vector2 velocity = _rigidbody.linearVelocity;
-        velocity.x = _moveInput * moveSpeed;
+        velocity.x = _inputController.MoveInput * moveSpeed;
         _rigidbody.linearVelocity = velocity;
 
-        if (_jumpRequested)
+        if (IsGrounded() && _inputController.JumpRequested)
         {
-            _jumpRequested = false;
+            _inputController.ConsumeJumpRequest();
             _rigidbody.linearVelocity = new Vector2(_rigidbody.linearVelocity.x, 0f);
             _rigidbody.AddForce(Vector2.up * jumpForce, ForceMode2D.Impulse);
         }
@@ -52,53 +42,4 @@ public class PlayerController : MonoBehaviour
 
         return Physics2D.OverlapCircle(groundCheck.position, groundCheckRadius, groundLayer);
     }
-
-    private float ReadHorizontalInput()
-    {
-        float keyboardInput = 0f;
-
-        if (Keyboard.current != null)
-        {
-            if (Keyboard.current.leftArrowKey.isPressed || Keyboard.current.aKey.isPressed)
-            {
-                keyboardInput -= 1f;
-            }
-
-            if (Keyboard.current.rightArrowKey.isPressed || Keyboard.current.dKey.isPressed)
-            {
-                keyboardInput += 1f;
-            }
-        }
-
-        float gamepadInput = 0f;
-
-        if (Gamepad.current != null)
-        {
-            gamepadInput = Gamepad.current.leftStick.ReadValue().x;
-        }
-
-        float input = Mathf.Abs(gamepadInput) > Mathf.Abs(keyboardInput) ? gamepadInput : keyboardInput;
-
-        return Mathf.Clamp(input, -1f, 1f);
-    }
-
-    private bool IsJumpTriggered()
-    {
-        bool jump = false;
-
-        if (Keyboard.current != null)
-        {
-            jump |= Keyboard.current.spaceKey.wasPressedThisFrame ||
-                    Keyboard.current.wKey.wasPressedThisFrame ||
-                    Keyboard.current.upArrowKey.wasPressedThisFrame;
-        }
-
-        if (Gamepad.current != null)
-        {
-            jump |= Gamepad.current.buttonSouth.wasPressedThisFrame;
-        }
-
-        return jump;
-    }
-
 }
