+++ Assets/Scripts/PlayerInputController.cs.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: d3aee7698f04b4d3b2dc59ed1a1e5283
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
